Add `BlutgangRpcMethod` and `EthRpcMethod` types
Adds better handling of raw string method types by placing an
enum in front of static string values. This slightly reduces
size due to less `&'static str` allocations and provides a more
idiomatic interface for de/serialization and removing the need for
`Clone` when deserialization is not necessary. This could also pave
the way for a small library for interacting with blutgang as the
interface changes, reducing the changes of invalid method calls.